Feat: refactor project structures
