+++ src/pages/Staking/StakeOptionModal/Manage/index.tsx
@@ -13,7 +13,8 @@ import {
   useTheme,
   useColorMode,
 } from '@chakra-ui/react';
-import {closeSale, fetchWithdrawPayload} from '../staking.reducer';
+import {fetchWithdrawPayload} from './utils/fetchWithdrawPayload';
+import {closeSale} from '../../actions';
 import {useWeb3React} from '@web3-react/core';
 import {useAppDispatch, useAppSelector} from 'hooks/useRedux';
 import {useCallback} from 'react';
@@ -27,7 +28,7 @@ export const ManageModal = () => {
   const theme = useTheme();
   const {colorMode} = useColorMode();
   let balance = data?.data?.stakeContractBalanceTon;
-  
+
   const withdrawPayload = async (data: any) => {
     const result = await fetchWithdrawPayload(
       data.library,
@@ -153,7 +154,6 @@ export const ManageModal = () => {
               color={'white.100'}
               fontSize={'0.750em'}
               fontWeight={100}
-              
               onClick={() =>
                 dispatch(openModal({type: 'stakeL2', data: data.data}))
               }
@@ -166,7 +166,6 @@ export const ManageModal = () => {
               color={'white.100'}
               fontSize={'12px'}
               fontWeight={100}
-              
               _hover={{backgroundColor: 'blue.100'}}
               onClick={() =>
                 dispatch(openModal({type: 'unstakeL2', data: data.data}))
@@ -179,7 +178,6 @@ export const ManageModal = () => {
               color={'white.100'}
               fontSize={'12px'}
               fontWeight={100}
-              
               _hover={{backgroundColor: 'blue.100'}}
               onClick={() => withdrawData('withdraw', data)}>
               Withdraw
@@ -190,7 +188,6 @@ export const ManageModal = () => {
               color={'white.100'}
               fontSize={'12px'}
               fontWeight={100}
-              
               _hover={{backgroundColor: 'blue.100'}}
               onClick={() =>
                 dispatch(openModal({type: 'swap', data: data.data}))
