+++ src/pages/Staking/StakeOptionModal/claim.tsx
@@ -12,7 +12,7 @@ import {
   useTheme,
   useColorMode,
 } from '@chakra-ui/react';
-import {claimReward} from '../staking.reducer';
+import {claimReward} from '../actions';
 import {useWeb3React} from '@web3-react/core';
 import {useAppDispatch, useAppSelector} from 'hooks/useRedux';
 import {closeModal, selectModalType} from 'store/modal.reducer';
@@ -54,7 +54,8 @@ export const ClaimOptionModal = () => {
               Claim
             </Heading>
             <Text color="gray.175" fontSize={'0.750em'} textAlign={'center'}>
-              You can claim {claimed?claimed:'0.00'} TOS and earned {earned?earned:'0.00'} TOS
+              You can claim {claimed ? claimed : '0.00'} TOS and earned{' '}
+              {earned ? earned : '0.00'} TOS
             </Text>
           </Box>
 
@@ -80,7 +81,7 @@ export const ClaimOptionModal = () => {
               _focus={{
                 borderWidth: 0,
               }}>
-              {claimed?claimed:'0.00'} TOS
+              {claimed ? claimed : '0.00'} TOS
             </Text>
           </Stack>
