+++ src/pages/Staking/StakeOptionModal/stakeInLayer2.tsx
@@ -17,7 +17,7 @@ import React, {useCallback, useState} from 'react';
 import {useWeb3React} from '@web3-react/core';
 import {useAppDispatch, useAppSelector} from 'hooks/useRedux';
 import {closeModal, selectModalType} from 'store/modal.reducer';
-import {stakeToLayer2} from '../staking.reducer';
+import {stakeL2} from '../actions';
 
 export const StakeInLayer2Modal = () => {
   const {account, library} = useWeb3React();
@@ -130,7 +130,7 @@ export const StakeInLayer2Modal = () => {
               color="white.100"
               fontSize="14px"
               onClick={() =>
-                stakeToLayer2({
+                stakeL2({
                   userAddress: account,
                   amount: value.toString(),
                   contractAddress: data?.data?.contractAddress,
