+++ src/pages/Staking/actions/index.ts
@@ -1,0 +1,19 @@
+import {stakePayToken} from './stakePayToken';
+import {unstake} from './unstake';
+import {claimReward} from './claim';
+import {unstakeL2} from './unstakeFromLayer2';
+import {withdraw} from './withdraw';
+import {swapWTONtoTOS} from './swap';
+import {closeSale} from './closeSale';
+import {stakeL2} from './stakeToLayer2';
+
+export {
+  stakePayToken,
+  unstake,
+  claimReward,
+  stakeL2,
+  unstakeL2,
+  withdraw,
+  swapWTONtoTOS,
+  closeSale,
+};
