+++ src/pages/Staking/actions/unstakeFromLayer2.ts
@@ -1,0 +1,41 @@
+import {getSigner, getRPC} from 'utils/contract';
+import {Contract} from '@ethersproject/contracts';
+import store from 'store';
+import {setTxPending} from 'store/tx.reducer';
+import {REACT_APP_TOKAMAK_LAYER2} from 'constants/index';
+import {utils} from 'ethers';
+import * as StakeTON from 'services/abis/StakeTON.json';
+
+type UnstakeFromLayer2 = {
+  userAddress: string | null | undefined;
+  amount: string;
+  contractAddress: string;
+  status: string;
+  library: any;
+  handleCloseModal: any;
+};
+
+const rpc = getRPC();
+
+export const unstakeL2 = async (args: UnstakeFromLayer2) => {
+  const {userAddress, amount, contractAddress, library} = args;
+  if (userAddress === null || userAddress === undefined) {
+    return;
+  }
+  const signer = getSigner(library, userAddress);
+  const StakeTONContract = new Contract(contractAddress, StakeTON.abi, rpc);
+  const wtonAmount = utils.parseUnits(amount, '27');
+  try {
+    const receipt = await StakeTONContract.connect(
+      signer,
+    ).tokamakRequestUnStaking(REACT_APP_TOKAMAK_LAYER2, wtonAmount);
+    store.dispatch(setTxPending({tx: true}));
+    alert(`Tx sent successfully! Tx hash is ${receipt.hash}`);
+    if (receipt) {
+      store.dispatch(setTxPending({tx: false}));
+    }
+  } catch (err) {
+    store.dispatch(setTxPending({tx: false}));
+    console.log(err);
+  }
+};
